Reference data for global metabolic construction

Reference data for global metabolic construction

The reference metabolic pathway data contains KOs, substrates and products, as well as a constructed reference global network, which used for metabolic network reconstruction

Format

The format is: List of 7 KO, substrate, product, user, date, version, reference network

Details

Information this dataset is involved:

KO, all KEGG orthlogy enties in KEGG metabolic pathways.

substrate, substrate of enzymatic reactions in all KEGG metabolic pathways.

product, product of enzymatic reactions in all KEGG metabolic pathways.

user who download this data.

date, the date this data is downloaded.

version, R version used to obtained it.

network, the global network which is reconstructed based on all the metabolites.

References

https://www.bioconductor.org/packages/release/bioc/html/mmnet.html