seedset-class

Object representing the seed sets of a given metabolic network

Slots

GsMN,

a igraph network

seeds,

a character list represents seeds of a given metabolic network which is composed of the KEGG compound index.

method

getGsMN, signature(object = "seedset"): get the genome scale metabolic network whose seed set is caculated

len, signature(object = "seedset"): return the number of source SCC

seedSize, signature(object = "seedset"): returns the sizes of each source SCCs

nonseed, signature(object = "seedset"): the non seeds of the GsMN

show, signature(object = "seedset"): show the short summary of a seedset class

confidencescore, signature(object = "seedset"): confidence score of the seed set

See also

getSeedSets,getGsMN,len, nonseed,seedSize,confidencescore

Examples

## Not run: ------------------------------------
# #' ## generate a metabolic network in igraph class and a seed set of this graph
# annodir <- system.file("extdata","koanno.tab",package = "RevEcoR")
# metabolic.data <- read.delim2(file=annodir,stringsAsFactors=FALSE)
# g <- reconstructGsMN(metabolic.data)
# seeds <- getSeedSets(g)@seeds
# seed.set <- new("seedset",GsMN = g, seeds = seeds)
## ---------------------------------------------